+++ src/components/common/ImagePlaceholder.jsx
