+++ src/components/layout/Layout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Navbar from "../Navbar.jsx";
+import Navbar from "../home/Navbar.jsx";
 
 export default function Layout() {
   return (
