+++ src/components/layout/LayoutAuth.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Navbar from "../Navbar.jsx";
+import Navbar from "../home/Navbar.jsx";
 import { useEffect } from "react";
 
 export default function Layout() {
