+++ src/components/layout/LayoutInstantiate.jsx
@@ -1,4 +1,4 @@
-import Sidebar from "../Sidebar.jsx";
+import Sidebar from "../common/Sidebar.jsx";
 import { Outlet } from "react-router-dom";
 
 export default function LayoutInstantiate() {
