+++ src/pages/EventMaker.jsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react';
 import { FaPalette, FaPuzzlePiece, FaEdit, FaBars, FaFileAlt } from 'react-icons/fa';
-import DashboardCard from '../components/DashboardCard';
+import DashboardCard from '../components/common/DashboardCard.jsx';
 import { useTranslation } from 'react-i18next';
 
 export default function EventMaker() {
