--- src/pages/ManualRegister.jsx
@@ -1,7 +1,0 @@
-export default function ManualRegister(){
-    return(
-        <div className="w-full min-h-svh p-8">
-            <h1 className="text-3xl font-bold">Manual Register</h1>
-        </div>
-    )
-}
