+++ src/pages/Users.jsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react";
 import { useUsers } from "../contexts/UsersContext";
 import { useKeycloak } from "@react-keycloak/web";
 import { useTranslation } from "react-i18next";
-import Pagination from "../components/Pagination.jsx";
-import CreateCard from "../components/CreateCard.jsx";
+import Pagination from "../components/common/Pagination.jsx";
+import CreateCard from "../components/common/CreateCard.jsx";
 import { FaUsers, FaUser, FaUsersCog, FaPlus, FaTrash, FaCheck, FaLock, FaShieldAlt, FaKey, FaExclamationTriangle } from "react-icons/fa";
 import KeycloakAdminService from "../services/KeycloakAdminService";
 import { UserExcelImport } from "../components/users/UserExcelImport";
