+++ src/pages/plugins/FloorPlans.jsx
@@ -7,9 +7,9 @@ import {
   FaUpload,
 } from "react-icons/fa";
 import { useKeycloak } from "@react-keycloak/web";
-import { axiosWithAuth } from "../utils/axiosWithAuth";
-import { baseUrl } from "../consts";
-import { useNotification } from "../contexts/NotificationContext";
+import { axiosWithAuth } from "../../utils/axiosWithAuth.js";
+import { baseUrl } from "../../consts.js";
+import { useNotification } from "../../contexts/NotificationContext.jsx";
 
 const apiUrl = `${baseUrl}/floor-plan-plugin/floor_plan`;
 const ITEMS_PER_PAGE = 6;
